+++ docs/.env.example
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_BUTTONDOWN_API_KEY=
+BUTTONDOWN_API_KEY=
